Make viewModel factories static, fix single apod
Part of manual DI is defining our own viewModel factories.
We had them as classes to be instantiated.

This change makes them already instantiated object
part of their viewModel, which makes it more intuitive.

Also (This should have been a part of another change but whatever),
the singleApod screen did not work because of bad parameter calls
to the savedStateHandle, which was fixed and the used savedStateHandle
was made reusable, so that multiple viewModels can share it.

Signed-off-by: Šimon Ukuš <xukus@stuba.sk>